+++ src/client/oauth.test.ts
@@ -3,7 +3,7 @@ import { OAuthClient } from './oAuth.js';
 import { describe, it, expect } from 'vitest';
 import fetchMock from 'fetch-mock';
 import { FetchLike, OAuthDb } from '../common/types.js';
-import { mockResourceServer, mockAuthorizationServer } from '../testHelpers.js';
+import { mockResourceServer, mockAuthorizationServer } from './clientTestHelpers.js';
 import { OAuthAuthenticationRequiredError } from './oAuth.js';
 
 function oauthClient(fetchFn: FetchLike, db?: OAuthDb, isPublic: boolean = false, strict: boolean = true, callbackUrl: string = 'https://example.com/mcp/callback') {
